Refactor command parsing by moving CommandsParser to CommandParameters.kt and removing the old CommandsParser file.
